[FEATURE] Solr AJAX Search Plugin
This adds a plugin-based interface to the new Solr AJAX Widget,
allowing the options the Widget supports to be configured through
plugin settings.

The new plugin is added as a new content element type.

In addition, some cleanups in the localconf/tables configuration
files has been done in order to optimize loading times and
enabling/disabling of features.

Change-Id: I99eeaa4d32d930055d0e7cd88c6ff42b1d1cff04