[BUGFIX] Insertion of new content after last element in FCE area
This fixes a case where tx_fed_fcecontentarea was being incorrectly
set when a new content element was added by creating a new content
element after another element if that element is inside an FCE's
content area. The fix consists of moving the detection of FCE area
when saving a record without an initial command (move) away from
postProcessing hook into the preProcessing hook instead.

Thanks to Frank Gerards for reporting and diggin!

Change-Id: I9238f48794ca3c5a5244d0763b20ed59eeb1ab90
Fixes: #32681