[FEATURE] Dynamic Backend Layout from Grid defined in Fluid Page template
This leverages a very recent change for the "cms" extension to get
a hook into the Backend Layout fetching routines, manipulating it
according to the Fluid Page template.

In you template register the Flux namespace and use
<flux:flexform.grid> ... </flux:flexform.grid> to define your grid.
Use Fluid conditions and loops - whatever you like - to make the
layout dynamic.

The "real" Backend Layout acts as a default value, preserving the
current behavior if no grid is defined in the template file, so
integration should be quite gentle.

If you're feeling lucky this is the change in TYPO3 necessary to
run this feature:

https://review.typo3.org/#change,7743

Change-Id: Iedda61c0fd437e447149ed48c9680a400d45a0f7