[TASK] Move definition of fallback page template to TypoScript
In addition this adds an extension configuration flag to disable
the fallback page template feature, allowing FED to fail if no
page template is selected.

If the fallback is enabled the integrator can define a fallback
template filename to use if no template is selected anywhere in
the rootline for any given page.

Change-Id: I935e0baf043f26cf652670052cfb1f2ca8c086e5
Resolves: #34119