[BUGFIX] Tx_Fed_Service_File::getUploadedFiles should check for empty filenames
When using Tx_Fed_Service_File::autoUpload Tx_Fed_Service_File::getUploadedFiles is called
which does not check if an filename in $_FILES is empty. This causes Tx_Fed_Service_File::move to be
called on an unexisting file.

I patched the method to look for unset or empty strings, this way non-mandatory file upload fields
could be used.

Change-Id: I5e49659afba44bd28e1364d8b465c422773c8b5f
Fixes: #34521
Releases: 4.6