[BUGFIX] Field name for fed:select bound to 1:1 property and multiple not specified
This fixes a case where the field name for fed:select would become
incorrect if the "multiple" attribute was not set (to something
equal to false) and the field is bound to a 1:1 property (i.e. a
single selection <select> field for a property.

The fix improves the checks employed to determine wether multiple
mode is desired/required.

Thanks to Stefano Cecere for report and solution!

Change-Id: I76902db795572a5558f21f242521b854836b434a
Fixes: #34955