[TASK] Move built-in Fluid Content Elements to EXT:fluidcontent
This change removes the templates and configuration used to provide
the built-in Fluid Content Elements. If installed, EXT:fluidcontent
will provide those elements in the future (with compatibility for
existing installs using the built-in elements; only the location
of the template files and where the setup comes from is changed).

Change-Id: Ie7762512631828506b113636cb53f96db7448ebf
Fixes: #36421