[BUGFIX] Allow Data/Var ViewHelper to set empty values
This adds a OR check which force-sets the value if a value argument
was used.

Old behavior: variable would be echoed if $value not set or empty/NULL

New behavior: variable will always be set if $value attribute is used
or child nodes exist and evaluate to a value not strictly NULL. If you
are setting a variable using ViewHelpers nested in fed:data.var and
these ViewHelpers are capable of returning NULL, instead use the value
attribute on fed:data.var and inline notation for your ViewHelper as
attribute value.

Change-Id: I974716d460984be7e158d2ea29c905f555c431ba
Fixes: #36460