[BUGFIX] Output from page:renderContent VH after tx_flux_parent
This fixes output issues where old content elements would not be
output to the FE after updating Flux, due to a missing check on
tx_flux_parent. However, to preserve compatibility with older
versions of Flux, the check is not done on tx_flux_parent but
extended to catch a second scenario on tx_flux_column IS NULL.

Thanks to Anders Gissel for the patch!

Change-Id: I44164fb8484ee0481dd313f74f2f8cd6ca339135
Fixes: #39381
Fixes: #39375