[FEATURE] Sliding of FlexForm values in page configurations
Note: this feature is enabled through extension configuration!

This change utilizes the new hook-based ConfigurationProvider methods
to operate on page records which contain a hierarchy. The processing
changes changes the previous behavior:

* When creating new pages, values come from first parent that has template
* When editing pages, children retain their own (possibly inherited) config
* When moving pages, all values of parent and child are preserved

Into a new behavior:

* When creating pages, values from first parent that has un-inherited config
* When updating page configuration, children with identical config also update
* Child updating stops if child page template is different
* Child updating stops if child page configuration is different
* If child updating is stopped, neither that child nor its children are touched
* Shortcuts are not followed but the shortcut itself gets updated if matched
* Moving a page considers and inherits on match the new parent's configuration

The matching is done recursively, unordered and in completely which means that
for example:

* Array('a' => Array('b', 'c'), 'd' => 1) == Array('d' => 1, 'a' => Array('c', 'b'))
* Array('a', 'b', 'c') == Array('c', 'b', 'a');
* Array('a', 'b', 'c') != Array('a', 'b', 'c', 'd');
* Array('a', 'b', 'c') != Array('a', 'b');

Please note that a weak typed comparison is performed but still respecting
completeness, which means that:

* Array(0, 1) == Array(NULL, TRUE)
* Array(0, 0) != Array(0, 0, 0)

You are free to change the order of your Flux fields in the template; children
which have an old data array containing the same values but in a different
order will still be recognized.

Last note: when changing the page template of a page to a value that is NOT
the same or empty (which means page uses a template different from the parent)
the stored FlexForm values are NOT truncated. This CAN cause issues with value
sliding if your template files do not use the same fields.

A future update should take care of this last issue with retained field values
from previously stored forms not being truncated. Ideally this should be possible
to set using a TS template option so it can be used on a per-page or per-tree
basis.

This greatly improves the situation described in #36550 but FED still will have
a tendency to lose the tie if page templates are changed, page is saved and
then changed back again. So don't experiment with templates on published pages.

Change-Id: Ib3ab18d651079cf90f9f9c07649289cfd5fb4a7d
Fixes: #36550