[TASK] Special icon for Fluid Content Elements in page and list modules
This change tricks the icon factory into loading custom icon for
tt_content records where CType is "fed_fce". The icon used is the
root page icon from the page tree (which coincidentally is the same
grayscale as the one chosen for FED and other Extbase extensions).

Thanks to Domi Garms for reporting!

Change-Id: I450330627854368472500395453552e8ec3967a4
Fixes: #39495