[TASK] Coding guidelines, missing/unused vars, deprecated classes etc.
This one is a huge cleanup of almost all usage of internally
deprecated classes and methods along with fixes for a huge
number of smaller issues such as missing PHPdoc annotations
or incorrectly documented parameter types.

The result is as expected a nice performance boost due to no
longer writing deprecation log warnings.

Set as fixing #34746 but with one exception; the DocumentHead
class which still needs to be in place for legacy purposes
but which will not generate any deprecation log warnings, just
IDE warnings.

Fixes: #34746
Fixes: #38953
Fixes: #39205
Fixes: #38908
Change-Id: Ic70fee1b20ae347681bccf51f2e9a9ce1d212398