[BUGFIX] Menu and BreadCrumb VH, full overlay usage and label field list argument
This change makes the Menu and BreadCrumb ViewHelpers use the full
record overlay when fetching page records. In addition a new argument
is added, replacing "useNavigationTitle" with a list of preferred and
prioritized field names to use as label for the generated links.

Default is "nav_title,title" which matches old behavior. Can be
modified to for example "tx_myext_titleoverride,nav_title,title" if
you have added tx_myext_titleoverride as a special field on pages.

The first field that has content will be used. The fields are overlayed
before attempting to resolve a label.

Thanks to Matthias Nitsch for reporting!

Change-Id: I78355a6f0003f6bf7fec3d012dd6846e94852c43
Fixes: #35779