[TASK] Get rid of extension configuration toggle enableFluidContentElements
This change...

* Removes extension configuration option enableFluidContentElements
* Always registers the CType for Fluid Content
* Always loads the most basic TypoScript needed to run FCEs

It is still possible to completely disable all Fluid Content Element
processing by simply unsetting the fed_fce CType in the tt_content
TypoScript location and unsetting plugin.tx_fed.fce. Or by simply
removing all extensions that provide content elements (for example
EXT:fluidcontent and EXT:fedsite).

This will also prevent issues with has-been child content elements
which would be rendered unintentionally (double) in FE even though
they did not appear in BE except in the list module.

The new behavior does add a bit of TCA that cannot easily be removed,
however, this bit is fairly small and is entered only for the CType
fed_fce which means that unsetting this CType through TypoScript
will effectively also ignore all the TCA that is added. Adding the
TCA itself requires very little processing time and of course
supports caching.

Set as fixing #34688 but this is of course only true once the intro
package itself is updated - or if FED is updated after installing it.

Change-Id: I3815f0748d504d8ae1d32c4948473e651b77aa23
Fixes: #34799
Fixes: #34688