[TASK] Allow PDF generation of URL instead of HTML source
This change introduces an "useHtml" argument for the PDF ViewHelper,
which switches between using HTML source and using the current URL
as source when generating the PDF file.

Note that the new default behavior is to use the URL. If you need
to generate PDFs of access protected or AJAX enabled pages, switch
on the "userHtml" argument (for those pages where it is needed).

Thanks to PowerKiKi for the patch!

Change-Id: I1949aeaf69ec804dd3f2b09202b1dc5460632485
Fixes: #35345