[FEATURE] Added 'insertJSInBody' argument to return uploader-code as HTML
If the uploader is loaded through an AJAX-call, the actual
instantiation and configuration code (residing in the header block)
will most likely be completely ignored. This option makes the
viewhelper return the actual script-block, so the browser can
instantiate the uploader asynchronously. The user will of course need
to include the javascript files himself, then.