[FEATURE] Added initFunctionName and storedValue-arguments; internal handling of stored values
Added arguments "initFunctionName", which can be used to replace the
standard jQuery-method used to initialize the uploader, and
"storedValue", which can be used to overrule the form field's own value
with preset values. The latter relies on a new internal function,
"getStoredValue()", which is in heavy need of some TLC to make it
support object storages.