[FEATURE] Realurl AutoConfigurationGenerator
This will make up the final implementation of the auto-route-builder.
The hook that post processes rules when automatic Realurl configuration
is enabled is leveraged to analyze the current site for all occurrences
of plugins which…

* Use arguments. Without this criteria it makes little sense to make
any rules. A future revision will allow routing to be triggered using
an annotation on each controller method, for example to allow pretty
URLs for Controller/action.
* Is placed at the top location of all Extbase plugins on the
particular page where the plugin is inserted. This is the only way
Realurl is capable of ensuring no colissions happen, but it does still
allow you to insert multiple plugins that use the same Realurl rule on
one page - if you are careful about allowing actions and making
arguments compatible.

This means that the realurl_autoconf.php file must be removed when the
plugin/page instances change before the URL routes will be rebuilt.

This first version supports lookups of Domain Objects by inspecting the
TCA of the Model and using the related table name and label field in a
generic configuration.