[TASK] Improve argument handling in Routing feature
This change further improves the handling of (optional) arguments by
refining both the Realurl rules which are created as well as the method
of retrieval. In addition two small caches have been implemented; one
which caches TCA that is loaded on-the-fly when generating the auto
configuration and another which allows loading an already generated
alias from the DB rather than creating a new alias. The first cache
improves performance when creating the automatic configuration and the
second cache improves general link creation performance on pages
(before they are cached, naturally) and always improves uncached
typoLink usage.