[BUGFIX] Avoid 6.0 crash on saving elements by avoiding cache clear
Internally this has the result that the TS for FCEs is not rebuilt at a time that is bad for the TYPO3 internals which are being queried to detect FCEs. The FCE detection is currently part of both EXT:fed and EXT:fluidcontent but will be rebuilt in EXT:fluidcontent, from the ground up. At that point the official recommendation is to not use FED unless combined with EXT:fluidcontent (and EXT:fluidpages if desired).

Fixes: https://github.com/NamelessCoder/fluidcontent/issues/10